refactor: 🔨 Keep all json process json in one place
